draft of safety ungrabbing logic for linux
If our input handling code cucked itself into a total system hardlock
where we are grabbing all of the keyboard but not passing any of input
further down, we need some kind of a safety net to ungrab devices if
this happens. This logic implements a separate thread and a timer that
sends a special secret keyboard code to a separate virtual keyboard,
and then checks if this code got received and acknowledged. If we
don't see it happening, we just ungrab everything we grabbed previously.